View Screen Working with Struct/String basic formatting
The View Screen now works with the structs, by converting their contents to a basic formatted string, and printing this with Qt's functionality to a basic viewing box. This is unformatted but sufficient.

WARNING: It only works with the Buttered_Toast recipe as of now, as search/selection functionality has not been updated to ignore use of File Explorer window.